+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **New Features:**
 
 - Support for MITRE ATT&CK Navigator heatmaps. (#11) (@fukusuket)
+- Added a `configure` command to configure Windows settings to various baselines. (#12) (@fukusuket)
 
 **Bug Fixes:**